+++ src/components/UI/atoms/Button/Button.module.scss
@@ -17,12 +17,6 @@
     line-height: normal;
   }
 
-  // &__content {
-  //   display: flex;
-  //   align-items: center;
-  //   gap: 10px;
-  // }
-
   &__icon {
     color: #1a324c;
   }
