--- src/components/UI/atoms/MegaMenuCard/MegaMenuCard.interface.ts
@@ -1,11 +1,0 @@
-export interface IMegaMenuCard {
-  title: string
-  logo?: string
-  icon?: string
-  className?: string
-  quotas: number
-  href?: string
-  target?: string
-  onClick: () => void
-  showRigthArrow?: boolean
-}
