+++ src/components/UI/atoms/index.ts
@@ -31,6 +31,4 @@ export * from './ToggleButton'
 export * from './UserMenuButtonAnalyst'
 export * from './ComparativeCounter'
 export * from './MegaMenuTab'
-export * from './MegaMenuPopover'
 export * from './Button'
-export * from './MegaMenuCard'
