+++ src/components/UI/molecules/MegaMenuCards/index.ts
@@ -1,1 +1,2 @@
 export { default as MegaMenuCards } from './MegaMenuCards.component'
+export * from './MegaMenuCards.interface'
