+++ src/components/UI/molecules/index.ts
@@ -61,3 +61,5 @@ export * from './UserMenuAnalyst'
 export * from './UserMenuWrapperAnalyst'
 export * from './MegaMenuJobsTabs'
 export * from './MegaMenuCards'
+export * from './MegaMenuCard'
+export * from './MegaMenuPopover'
