--- src/components/UI/organism/MegaMenuMainHeader/MegaMenuMainHeader.interface.ts
@@ -1,6 +1,0 @@
-import { IMegaMenuCards } from '@components/UI/molecules/MegaMenuCards/MegaMenuCards.interface'
-
-export interface IMegaMenuMainHeader {
-  prueba: string
-  jobs: IMegaMenuCards[]
-}
