+++ src/components/UI/organism/MegaMenuMainHeader/index.ts
@@ -1,2 +1,1 @@
-export * from './MegaMenuMainHeader.interface'
 export { MegaMenuMainHeader } from './MegaMenuMainHeader.component'
