+++ src/components/UI/template/MegaMenu/MegaMenu.component.tsx
@@ -3,8 +3,7 @@ import { MegaMenuSocialHeader } from '@components/UI/organism/MegaMenuSocialHead
 import { MegaMenuProvider } from './MegaMenu.context'
 import { IMegaMenuContext } from './MegaMenuContext.interface'
 import { MegaMenuMainHeader } from '@components/UI/organism/MegaMenuMainHeader/MegaMenuMainHeader.component'
-import { MegaMenuCard } from '@components/UI/atoms'
-import { Breadcrumbs, MegaMenuCards } from '@components/UI/molecules'
+import { Breadcrumbs, MegaMenuCard, MegaMenuCards } from '@components/UI/molecules'
 import styles from './MegaMenu.module.scss'
 import { IMegaMenu } from './MegaMenu.interface'
 import { withMegaMenuContainer } from '@components/hoc'
